Create README.md
John David Conley
700645761
	ReadMe:
The execution here is simple, simply run the code in the .py file and all the problems will go in sequence. For the most part, the only comments are the directions, as these are for the most part fairly straightforward, and most other notices are instead printed. Occasionally, it asks for numeric inputs. The visuals for #10 are included at the end of the Word edition of the source code file, after the sample output.

I had to upload two videos because the visuals broke on the first one, but I explained it better there than on the second one due to needing to monitor the second one more to prevent similar errors.
Video 1: https://youtu.be/eXHMguVV8oE
Video 2: https://youtu.be/e_5WtHWs9xE 
GGitHub repo: https://github.com/DavidConley/machinelearning01